requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
python-dotenv==1.0.0
openai==1.23.2
SpeechRecognition
scikit-learn 
numpy
gtts
bs4
langchain_openai 
langchain_core
langchain-community 
langchain
azure-cosmos
azure-search-documents
azure-identity
requests
flask-sse
quart
gTTS
pyttsx3
langdetect
azure-cognitiveservices-speech
# pip install 'flask[async]'
# quart --upgrade